Propose COMMUNICATION.md as a way to improve vibes
From the document itself:

OpenTelemetry has a Code of Conduct which primarily pertains to
"bannable" offenses: things like harassment, personal attacks, and so
on. Certainly OpenTelemetry does not tolerate any such behavior.

And while we all deserve to collaborate without fear of overt harassment
and personal attacks, that is a very low bar! This document is more
aspirational and has fewer "teeth" (i.e., it's not about "banning"
people), but it's also very important: it describes how we wish to
communicate with each other and how we aspire to create an environment
that cultivates trust, even in the face of inevitable technical and
organizational disagreement.