Specify the steps in the OTel donation process [take two] (#831)
* Specify the steps in the donation process

Most of these were previously understood but not properly documented. The step about the marketing guidelines is meant to be consistent with the GC discussion at our meetings over the past two weeks.

* fix indentation

* Respond to PR comments

Many of these are adapted from Sergey's suggestions in the GitHub PR
flow. Due to branch protections they could not be applied automatically.

* Add an issue template

* Respond to review comments

* Line wrapping