Add Community Manager/Maintainer Role
This PR adds a new project-level maintainer role, 'Community Manager'.

## Problem Statement
Over the past several years, we've seen explosive growth of the OpenTelemetry community across many fronts. Individual and corporate contributors have more than doubled in count, and contributions have followed suit. As the project matures, so too have the needs of the auxiliary SIGs and working groups. This can be seen in the shifting role of the Web SIG, for example -- originally simply responsible for website maintenance, it has matured into a 'communications' SIG responsible for general cross-functional events, the project blog, social comms, documentation, tutorials, and much more. 

This split in focus has resulted in a sub-optimal level of focus for some critical project needs (documentation being a particularly poignant example), as well as unclear ownership for certain channels (social, events). To resolve this split in focus, I propose a new 'community manager' role that can provide accountability and ownership for these disparate, but related, functions.

## Responsibilities and Portfolio
Some examples of this role's responsibilities include the following:
- Working with the CNCF to plan and organize OpenTelemetry-specific events such as OpenTelemetry Community Day
- Managing the OpenTelemetry blog and social media presence (creating content, ensuring quality of posted content, etc.)
- Working with the CNCF and community organizations on cross-functional projects (such as a contributor's fund)
- Reporting and tracking community health for the GC

In general, the objective of this role is to have a single point of contact to manage wide-ranging community initiatives and ensuring their successful execution. 

## Q&A
_"How is this different from Communications SIG lead?"_
Good question! This role is designed to codify the 'non-web/docs' functions of the Comms SIG lead into a new role, allowing the comms SIG to refocus on the website and project-wide documentation. The intent of this role is to closely collaborate with comms (as well as any other community-focused SIG/WG leads) to ensure alignment and resources are available to everyone.

_"How are CM's designated/how long do they serve?"_
The Community Manager reports to the Governance Committee and serves at their pleasure. More than one CM can be appointed by the GC, if desired/necessary. There's no specified length of service, but an annual review is suggested.

_"Isn't the GC/Comms SIG responsible for this stuff already?"_
Yes, technically -- that said, the GC is busy with other governance duties. This role codifies a position that already implicitly exists, while also allowing greater autonomy for the individual who holds it. For example, only project-level maintainers have access to CNCF Service Desk; This causes friction with the scope of Comms SIG maintainers when trying to perform actions that need access to that platform. 

Please let me know if you have any further questions/points of discussion.

cc @mtwo @bhs